% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-patchwork.R
\name{-.ggcall}
\alias{-.ggcall}
\alias{/.ggcall}
\alias{|.ggcall}
\alias{*.ggcall}
\alias{&.ggcall}
\title{Minus Operator for ggcall Objects}
\usage{
\method{-}{ggcall}(e1, e2)

\method{/}{ggcall}(e1, e2)

\method{|}{ggcall}(e1, e2)

\method{*}{ggcall}(e1, e2)

\method{&}{ggcall}(e1, e2)
}
\arguments{
\item{e1}{The left-hand side \code{ggcall} object.}

\item{e2}{The right-hand side \code{ggcall} object.}
}
\value{
A combined \code{ggcall} object after applying the minus operation.

A combined \code{ggcall} object after applying the division operation.

A combined \code{ggcall} object after applying the or operation.

A combined \code{ggcall} object after applying the multiplication operation.

A combined \code{ggcall} object after applying the and operation.
}
\description{
Applies the minus operator to \code{ggcall} objects, utilizing the corresponding operator from the \code{patchwork} package.

Applies the division operator to \code{ggcall} objects, utilizing the corresponding operator from the \code{patchwork} package.

Applies the or operator to \code{ggcall} objects, utilizing the corresponding operator from the \code{patchwork} package.

Applies the multiplication operator to \code{ggcall} objects, utilizing the corresponding operator from the \code{patchwork} package.

Applies the and operator to \code{ggcall} objects, utilizing the corresponding operator from the \code{patchwork} package.
}
\details{
This function allows for the subtraction of \code{ggcall} objects using the \code{patchwork} syntax.

This function allows for the division of \code{ggcall} objects using the \code{patchwork} syntax.

This function allows for the combination of \code{ggcall} objects using the \code{patchwork} syntax.

This function allows for the multiplication of \code{ggcall} objects using the \code{patchwork} syntax.

This function allows for the combination of \code{ggcall} objects using the \code{patchwork} syntax.
}
\keyword{internal}
